Restructure mutable-shape bqm tests to be more generic
